# Conflicts:
#	project/app/src/main/java/no/uio/ifi/in2000/project/data/alerts/MetAlertsDataSource.kt

+++ project/app/src/main/java/no/uio/ifi/in2000/project/model/forecast/LocationForecastResponse.kt
@@ -13,7 +13,7 @@ data class LocationForecastResponse(
 @Serializable
 data class Geometry(
     val type: String,
-    val coordinates: List<Int>
+    val coordinates: List<Float>
 )
 //Comments Comments Comments
 @Serializable
